refactor: API型定義を独立ファイルに分離
- frontend/types/api.ts を新規作成
- User, Book, UserBook, BookshelfBook, Bookshelf, Stats 型を移動
- lib/api.ts で型定義をインポートに変更
- 後方互換性のため re-export を維持
